requirements.txt
requests==2.31.0
schedule==1.2.0
pyyaml==6.0.1
python-dotenv==1.0.0
redis==4.6.0
celery==5.3.4
sqlalchemy==2.0.23
click==8.1.7
python-dateutil==2.8.2
selenium==4.15.0